+++ src/comms/SocketManager.ts
